wip: moved files to its own dedicated folder and had minor changes to certain files
