+++ cgi-bin/test.cgi
@@ -1,0 +1,5 @@
+#!/usr/bin/python3
+
+import os
+
+print("CGI Script Executed Successfully!")
