+++ src/main.cpp
@@ -10,8 +10,11 @@ void print_request(Request &request) {
 	}
 }
 
-int main(void)
+int main(int argc, char *argv[], char *env[])
 {
+	(void) argc;
+	(void) argv;
+	(void) env;
 	std::cout << "----------Request-----------" << std::endl;
 	Request request("GET", "/", "HTTP/1.1");
 	request.addHeader("Host", "localhost");
@@ -21,4 +24,8 @@ int main(void)
 	std::cout << "\n----------Response-----------" << std::endl;
 	Response &response = handle_get_response(request);
 	std::cout << response.toString();
+	delete &response;
+
+	// std::cout << "----------CGI TEST-----------" << std::endl;
+	// execute("./cgi-bin/test.cgi", env);
 }
